refactor: move crosshair to types/xy_chart/crosshair
