+++ src/types/xy_chart/crosshair/crosshair_utils.linear_snap.test.ts
@@ -1,10 +1,10 @@
-import { computeXScale } from '../types/xy_chart/utils/scales';
-import { BasicSeriesSpec } from '../lib/series/specs';
-import { Dimensions } from '../utils/dimensions';
-import { getGroupId, getSpecId } from '../utils/ids';
-import { ScaleType } from '../utils/scales/scales';
+import { computeXScale } from '../utils/scales';
+import { BasicSeriesSpec } from '../../../lib/series/specs';
+import { Dimensions } from '../../../utils/dimensions';
+import { getGroupId, getSpecId } from '../../../utils/ids';
+import { ScaleType } from '../../../utils/scales/scales';
 import { getCursorBandPosition, getSnapPosition } from './crosshair_utils';
-import { computeSeriesDomains } from './utils';
+import { computeSeriesDomains } from '../../../state/utils';
 
 describe('Crosshair utils linear scale', () => {
   const barSeries1SpecId = getSpecId('barSeries1');
